add layout for modal parallel route
