+++ packages/app/src/app/race/layout.tsx
@@ -1,0 +1,16 @@
+import * as React from "react";
+
+export default async function Layout({
+  children,
+  modal,
+}: {
+  children: React.ReactNode;
+  modal: React.ReactNode;
+}) {
+  return (
+    <div>
+      {children} 
+      {modal}
+    </div>
+  );
+}
